+++ b300_core/live_session.py
@@ -16,8 +16,6 @@ from .live_service import LiveMonitorService
 from .models import ProbeRef
 from .offline_symbols import OfflineSymbolTable
 from .ssh_debug_tunnel import find_available_loopback_port
-from .ssh_host_trust import trusted_known_hosts_file
-from .ssh_identity import managed_identity_file
 from .ssh_live_tunnel import SshLiveTunnel, SshLiveTunnelConfig
 from .tcl_client import SafeTclClient, TclEndpoint
 
@@ -51,6 +49,7 @@ class ClientLiveMonitorConfig:
     gateway_tcl_port: int = 6666
     symbol_roots: Tuple[Path, ...] = ()
     symbol_max_files: int = 128
+    show_console: bool = False
 
     def validate(self) -> None:
         if self.symbols is not None:
@@ -66,7 +65,8 @@ class ClientLiveMonitorConfig:
         SshLiveTunnelConfig(
             host=self.host, user=self.user, ssh_port=self.ssh_port,
             local_tcl_port=self.preferred_local_tcl_port, gateway_tcl_port=self.gateway_tcl_port,
-        ).validate()
+            show_console=self.show_console,
+        ).validate_endpoint()
 
 
 @dataclass(frozen=True)
@@ -198,8 +198,7 @@ class LiveMonitorSession:
         tunnel_config = SshLiveTunnelConfig(
             host=config.host, user=config.user, ssh_port=config.ssh_port,
             local_tcl_port=local_tcl, gateway_tcl_port=config.gateway_tcl_port,
-            identity_file=managed_identity_file(),
-            known_hosts_file=trusted_known_hosts_file(config.host, config.ssh_port),
+            show_console=config.show_console,
         )
         tunnel = self._tunnel_factory(tunnel_config)
         try: